+++ errors/InvalidObjectIdError.js
@@ -1,0 +1,11 @@
+const { errors } = require('../helpers/errorMessages');
+
+class InvalidObjectIdError extends Error {
+  constructor(docType) {
+    super();
+    this.statusCode = 409;
+    this.message = errors.objectId[docType];
+  }
+}
+
+module.exports = InvalidObjectIdError;
