+++ helpers/errorMessages.js
@@ -1,0 +1,21 @@
+module.exports.errors = {
+  invalidInput: {
+    name: 'Ошибка в поле Name.',
+    email: 'Ошибка в поле Email.',
+    about: 'Ошибка в поле About.',
+    avatar: 'Проблема с аватаркой.',
+    link: 'Проблема с изображением.',
+  },
+  invalidCredentials: 'Неправильные почта или пароль',
+  missingCredentials: 'Введите логин и пароль',
+  docNotFound: {
+    user: 'Такого пользователя нет',
+    card: 'Такой карточки нет',
+  },
+  emailInUse: 'Этот адрес электронной почты уже используется',
+  badPassword: (pswlength) => `Введите пароль длиной не менее ${pswlength} зн., состоящий из латинских букв, цифр и специальных символов`,
+  objectId: {
+    user: 'Ошибка в идентификаторе пользователя',
+    card: 'Ошибка в идентификаторе карточки',
+  },
+};
