+++ helpers/helpers.js
@@ -1,27 +1,7 @@
 const mongoose = require('mongoose');
+const { errors } = require('./errorMessages');
 const User = require('../models/user');
-
-const errors = {
-  invalidInput: {
-    name: 'Ошибка в поле Name.',
-    email: 'Ошибка в поле Email.',
-    about: 'Ошибка в поле About.',
-    avatar: 'Проблема с аватаркой.',
-    link: 'Проблема с изображением.',
-  },
-  invalidCredentials: 'Неправильные почта или пароль',
-  missingCredentials: 'Введите логин и пароль',
-  docNotFound: {
-    user: 'Такого пользователя нет',
-    card: 'Такой карточки нет',
-  },
-  emailInUse: 'Этот адрес электронной почты уже используется',
-  badPassword: (pswlength) => `Введите пароль длиной не менее ${pswlength} зн., состоящий из латинских букв, цифр и специальных символов`,
-  objectId: {
-    user: 'Ошибка в идентификаторе пользователя',
-    card: 'Ошибка в идентификаторе карточки',
-  },
-};
+const InvalidObjectIdError = require('../errors/InvalidObjectIdError');
 
 const passwordRegexp = /[\u0023-\u0126]+/;
 
@@ -48,9 +28,7 @@ function isUserExistent(id) {
 
 function isObjectIdValid(id, docType) {
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    const error = new Error();
-    error.docType = docType;
-    throw error;
+    throw new InvalidObjectIdError(docType);
   }
 }
 
@@ -85,12 +63,6 @@ function createDocHandler(promise, req, res, docType) {
     });
 }
 
-function getAllDocsHandler(promise, req, res, next) {
-  promise
-    .then((respObj) => res.send(respObj))
-    .catch(next);
-}
-
 function getLikeDeleteHandler(promise, req, res, docType, userId) {
   promise
     .orFail()
@@ -110,25 +82,9 @@ function getLikeDeleteHandler(promise, req, res, docType, userId) {
     });
 }
 
-function updateHandler(promise, req, res) {
-  promise
-    .orFail()
-    .then((respObj) => res.send(respObj))
-    .catch((err) => {
-      if (err instanceof mongoose.Error.DocumentNotFoundError) {
-        res.status(404).send({ message: `${errors.docNotFound.user}` });
-      } else if (err instanceof mongoose.Error.ValidationError) {
-        res.status(400).send({ message: joinErrorMessages(err) });
-      }
-    });
-}
-
 module.exports = {
   createDocHandler,
-  getAllDocsHandler,
   getLikeDeleteHandler,
-  updateHandler,
-  errors,
   joinErrorMessages,
   isUserExistent,
   isObjectIdValid,
